requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
fastapi>=0.111.0
uvicorn>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
psycopg2-binary
passlib
bcrypt
python-dotenv
